requirements.txt

asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.1
cssselect==1.2.0
Django==3.2.5
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
numpy
oauthlib==3.2.2
packaging==23.0
pandas
parsel==1.7.0
Pillow==9.4.0
Protego==0.2.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
queuelib==1.6.2
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.8.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
textblob==0.17.1
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
Twisted==22.10.0
twython==3.9.1
typing_extensions==4.4.0
urllib3==1.26.14
w3lib==2.1.1
zope.interface==5.5.2
